Move Country_Helper under Helpers namespace
